pyproject.toml
[project]
name = "metaseed"
version = "0.1.0"
description = "Schema-driven API for MIAPPE-compliant phenotyping metadata"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "sdrwacker" }]
keywords = ["miappe", "phenotyping", "metadata", "ontology", "ppeo"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "fastapi>=0.110",
    "typer>=0.12",
    "uvicorn>=0.29",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "openpyxl>=3.1",
    "mkdocs-material>=9.7.6",
    "isatools>=0.14.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "pre-commit>=3.7",
    "httpx>=0.27",
    "jupyter>=1.0",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "pymdown-extensions>=10.8",
]

[project.scripts]
metaseed = "metaseed.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/metaseed"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "SIM117", # nested with statements (normal for declarative UI code)
]

[tool.ruff.lint.isort]
known-first-party = ["metaseed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
markers = [
    "ui: marks tests as UI tests",
]

[tool.coverage.run]
source = ["src/metaseed"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pre-commit>=4.5.1",
    "pytest-asyncio>=1.3.0",
    "selenium>=4.20",
]
